Project/requirements.txt
appdirs==1.4.2
Django==1.11.29
djangorestframework==3.9.1
packaging==16.8
pyparsing==2.1.10
six==1.10.0
requests